pyproject.toml
[project]
name = "ai-companion2"
version = "0.1.0"
description = "AI gameplay companion coach"
authors = [
    {name = "Nathan Meek",email = "Nathan.meek4@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "mss (>=10.1.0,<11.0.0)",
    "pyside6 (>=6.10.0,<7.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "openai (>=2.8.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
